openrc-init: use poll for waiting on the fifo
this sets us up for being able to use "self-pipe trick" to handle
signals safely.

+++ src/openrc-init/openrc-init.c
@@ -31,6 +31,8 @@
 #include <sys/reboot.h>
 #include <sys/types.h>
 #include <sys/wait.h>
+#include <fcntl.h>
+#include <poll.h>
 #include <utmp.h>
 
 #ifdef HAVE_SELINUX
@@ -226,8 +228,7 @@ int main(int argc, char **argv)
 {
 	char *default_runlevel;
 	char buf[2048];
-	int count;
-	FILE *fifo;
+	int count, fifo;
 	bool reexec = false;
 	sigset_t signals;
 	struct sigaction sa;
@@ -300,33 +301,38 @@ int main(int argc, char **argv)
 
 	if (mkfifo(RC_INIT_FIFO, 0600) == -1 && errno != EEXIST)
 		perror("mkfifo");
+	fifo = open(RC_INIT_FIFO, O_RDONLY | O_NONBLOCK | O_CLOEXEC);
+	if (fifo == -1)
+		perror("open");
 
 	for (;;) {
-		/* This will block until a command is sent down the pipe... */
-		fifo = fopen(RC_INIT_FIFO, "r");
-		if (!fifo) {
-			if (errno != EINTR)
-				perror("fopen");
-			continue;
-		}
-		count = fread(buf, 1, sizeof(buf) - 1, fifo);
-		buf[count] = 0;
-		fclose(fifo);
-		printf("PID1: Received \"%s\" from FIFO...\n", buf);
-		if (strcmp(buf, "halt") == 0)
-			handle_shutdown("shutdown", RB_HALT_SYSTEM);
-		else if (strcmp(buf, "kexec") == 0)
-			handle_shutdown("reboot", RB_KEXEC);
-		else if (strcmp(buf, "poweroff") == 0)
-			handle_shutdown("shutdown", RB_POWER_OFF);
-		else if (strcmp(buf, "reboot") == 0)
-			handle_shutdown("reboot", RB_AUTOBOOT);
-		else if (strcmp(buf, "reexec") == 0)
-			handle_reexec(argv[0]);
-		else if (strcmp(buf, "single") == 0) {
-			handle_single();
-			open_shell();
-			init(default_runlevel);
+		enum { FD_FIFO, FD_COUNT };
+		struct pollfd pfd[] = {
+			[FD_FIFO] = { .fd = fifo, .events = POLLIN },
+		};
+
+		poll(pfd, FD_COUNT, -1);
+
+		if (pfd[FD_FIFO].revents & POLLIN) {
+			count = read(fifo, buf, sizeof(buf) - 1);
+			buf[count] = 0;
+			printf("PID1: Received \"%s\" from FIFO...\n", buf);
+
+			if (strcmp(buf, "halt") == 0)
+				handle_shutdown("shutdown", RB_HALT_SYSTEM);
+			else if (strcmp(buf, "kexec") == 0)
+				handle_shutdown("reboot", RB_KEXEC);
+			else if (strcmp(buf, "poweroff") == 0)
+				handle_shutdown("shutdown", RB_POWER_OFF);
+			else if (strcmp(buf, "reboot") == 0)
+				handle_shutdown("reboot", RB_AUTOBOOT);
+			else if (strcmp(buf, "reexec") == 0)
+				handle_reexec(argv[0]);
+			else if (strcmp(buf, "single") == 0) {
+				handle_single();
+				open_shell();
+				init(default_runlevel);
+			}
 		}
 	}
 	return 0;
